moved is50() to hubConfig, fixed img path for 5.0
