+++ classes/Table/class.hubAbstractTableGUI.php
@@ -69,7 +69,7 @@ abstract class hubAbstractTableGUI extends ilTable2GUI {
             $this->initStandardTableColumns();
         }
         if ($this->initTableRowTemplate() === false) {
-            if(hub::is50()){
+            if(hubConfig::is50()){
                 $this->setRowTemplate('tpl.std_row_template.html', 'Services/ActiveRecord');
             }else{
                 $this->setRowTemplate('tpl.std_row_template.html', 'Customizing/global/plugins/Libraries/ActiveRecord');
